+++ src/components/Carousel.js
@@ -1,10 +1,11 @@
 import React,{ useEffect, useState } from 'react';
 import HeaderCarousel from './HeaderCarousel';
-import MoviesCarousel from '../MoviesCarousel';
+import MoviesCarousel from './MoviesCarousel';
 const API_PATH = 'https://api.themoviedb.org/3/';
+import useFetch from '../utils/useFetch';
+
 
 export default function Carousel({title, category, mediaType, background, apiKey}) {
-
     // The media categories that can be found depending on the category of it.
     const mediaOptions = {
         trending: [
@@ -32,21 +33,12 @@ export default function Carousel({title, category, mediaType, background, apiKey
             }
         ]
     }
+    
     // It load the options that it can be selected in the carousel to feature different categories
     const [switchRoute, setSwitchRoute] = useState(mediaOptions[category]);
-    
-    // It stores the carousel data movie fetched from the API, it stores singular data depended on what you want display.
-    const [data, setData] = useState(undefined);
-    
-    // Fetches the data and then store that data.
-    const fetchData = () => {
-        const index = switchRoute.findIndex((element) => element.isSelected);
-        fetch(switchRoute[index].path)
-        .then(response => response.json())
-        .then(data => setData(data));
-    }
 
-    useEffect(() => fetchData(), [switchRoute]);
+    // Custom Hook 
+    const data = useFetch(switchRoute);
 
     return (
         <div className='carousel--container'>
