+++ src/components/PosterMovie.js
@@ -1,12 +1,11 @@
 import React from 'react';
-const IMAGE_PATH = 'https://image.tmdb.org/t/p/w220_and_h330_face/';
-
+import posterUrl from '../utils/posterUrl';
 export default function PosterMovie({score, posterPath}) {
     return (
         <div className={`image--container image--container--${score}`}>
             <img 
                 className='item--list--image'
-                src={`${IMAGE_PATH}${posterPath}`}
+                src={`${posterUrl}${posterPath}`}
             />
         </div>
     );
