+++ src/styles/_progressBar.scss
@@ -1,0 +1,37 @@
+// Custom Progress bar 
+.progress--bar {
+    max-width: 100px;
+    padding: 10px 7px;
+    border-radius: 50%;
+    border-style: solid;
+    background-color: $background-score;
+    color: $text-color2;
+    border-width: 3px;
+    position: absolute;
+    bottom: 90px;
+    left: 10px;
+    box-shadow: 3px 6px 7px $black;
+
+    &--high {
+        border-color: $high-score;
+    }
+
+    &--medium {
+        border-color: $medium-score;
+    }
+
+    &--low {
+        border-color: $low-score;
+    }
+}
+
+.progress--bar--average {
+    @include flexContainer(row, center, center);
+    @include fontSettings('Orbitron', .9em, medium, normal);
+    width: 27px;
+    max-width: 27px;
+}
+
+.percent--sign {
+    font-size: .8em;
+}
