+++ src/utils/apiKey.js
@@ -1,2 +1,2 @@
 const apiKey = '7c4a986b16b3b892bd7111a358d63e05';
-module.exports = apiKey;
+export default apiKey;
