+++ src/utils/posterUrl.js
@@ -1,0 +1,2 @@
+const posterUrl = 'https://image.tmdb.org/t/p/w220_and_h330_face/';
+export default posterUrl;
